introduce chemEntityType/SubType
Now only Bond, Ring, or Atom can be active.

To key properties/varaibles are used instead of _bondtype, _atomtype.
Now _chemEntityType hold information whether a ring, bond or atom type is selcted. Based on that the action is picked given a click.

The _chemEntitySubType, hold information on which exact bond, atom or ring has been selected